chore(xplr) break node string repr into generic function
